Few changes for jcommander Change version

+++ src/main/java/com/opencsi/jscepcli/KeyManager.java
@@ -19,7 +19,7 @@ public class KeyManager {
         
         try {
             kpg = KeyPairGenerator.getInstance("RSA");
-            kpg.initialize(keySize);
+            kpg.initialize(keySize.intValue());
             kp = kpg.genKeyPair();
             
         } catch (Exception e) {
